Safari-609.1.11.1/LayoutTests/fast/visual-viewport/ios/fixed-element-on-bottom-with-keyboard-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
The initial layout viewport is: [0 0 ; 320 548]
The layout viewport after showing the keyboard is: [0 0 ; 320 548]
The layout viewport after scrolling to the end is: [0 3452 ; 320 548]
The layout viewport after scrolling past the end is: [0 3452 ; 320 548]
PASS initialLayoutViewport.height is layoutViewportAfterShowingKeyboard.height
PASS layoutViewportAfterScrollingPastEnd.top is layoutViewportAfterScrollingToEnd.top
